moved main axios call to the view rather than at component level
